+++ earth/cpu_intr.c
@@ -7,25 +7,42 @@
 
 #include "egos.h"
 
-/* These are two static variables storing
- * the addresses of the handler functions;
- * Initially, both variables are NULL */
+static ulonglong mtime_get() {
+    uint low, high;
+    do {
+        high = REGW(0x200BFF8, 4);
+        low  = REGW(0x200BFF8, 0);
+    }  while ( REGW(0x200BFF8, 4) != high );
+
+    return (((ulonglong)high) << 32) | low;
+}
+
+static int mtimecmp_set(ulonglong time, uint core_id) {
+    REGW(0x2004000, core_id * 8 + 4) = 0xFFFFFFFF;
+    REGW(0x2004000, core_id * 8 + 0) = (uint)time;
+    REGW(0x2004000, core_id * 8 + 4) = (uint)(time >> 32);
+
+    return 0;
+}
+
+static uint QUANTUM;
+int timer_reset(uint core_id) { return mtimecmp_set(mtime_get() + QUANTUM, core_id); }
+
 static void (*kernel_entry)(uint, uint);
-int kernel_entry_init(void (*new_entry)(uint, uint)) {
-    kernel_entry = new_entry;
+int kernel_entry_init(void (*new_entry)(uint, uint)) { kernel_entry = new_entry; }
+void trap_entry(uint mcause) {
+    kernel_entry(mcause & (1 << 31), mcause & 0x3FF);
 }
 
 /* Both trap functions are defined in earth.S */
 void trap_from_M_mode();
 void trap_from_S_mode();
 
-void trap_entry() {
-    uint mcause;
-    asm("csrr %0, mcause" : "=r"(mcause));
-    kernel_entry(mcause & (1 << 31), mcause & 0x3FF);
-}
+void intr_init(uint core_id) {
+    QUANTUM = (earth->platform == ARTY)? 5000 : 500000;
+    mtimecmp_set(0x0FFFFFFFFFFFFFFFUL, core_id);
 
-void intr_init() {
+    earth->timer_reset = timer_reset;
     earth->kernel_entry_init = kernel_entry_init;
 
     /* Setup the interrupt/exception entry function */
